Updated Project to RC7.
Major issues during the migration to RC7. Had to step through a lot of
code to see what needed to be cleaned up.

Removed my Udemy tutorial section as it kept failing for some unknown
reason. As well, I'm not 100% sure I can reproduce the steps to get back
to RC7 so hopefully this is set up correctly. I do know I got some info
from the following posts:

http://stackoverflow.com/questions/38954205/migrating-from-router-deprecated-2-0-0-rc-2-to-router3-0-0-rc-1
(indicates that there is a bug with multi-line commenting. Hence using
/** on the spec.ts test files)