Hooked up Web Service to Login Component
I finally was abel to get the login component to check the web service.
Two things: You need to submit good credentials twice. I believe this is
because I need to better unerstand observables and how to make sure that
the request completes before continuing. Secondly, the API is being case
sensitive for Username. I'll fix that shortly. Good credentials are:
ADMIN, pword