Understanding Observables
Created a new view on the site to practice observables. Also created a
test class on the API with a lot of data to see how Observables behave
with a lot of data being downloaded. This can be found under Programming
Examples. I now need to apply this to the login component.