Working on Observables and Promises
Was able to figure out how to do an observable / promise to create a
clock at the top of the screen. Used that knowledge and fixed the login
function to respond after the click properly. Need to clean up the code
because it's a mess after all my testing. Also need to review AuthGuards
lol!