Fixed Login issue that required two submissions
Made changes to the login functions so that it didn't require two
correct submissions when logging in. I've noted the URL where I found
the fix re: chaining steps in the validateCredentials function.

This build shows credentials and some other logging in the debugger when
submitting credentials. As well, I want to separate the code better in
the next pass through this code.