Trying to figure out multiple nests in navigation tree
Was able to get breadcrumbs to work on first level but child after first
level requires an additional router-outlet. Because of this, I need to
figure out a proper implementation, which I will test with all of these
programming views I've made for testing. I'll nest this accordingly so I
can figure out how to set up things up to handle the breadcrumbes
properly.