Removed Quick Edit in List Tags
Removed the Quick Edit function in List Tags as it's a duplicate
funciton of Full Edit. Will use that implementation instead and have put
the correct URL in the list to go to the correct ID.

Edit Tag however, only brings up the correct ID, the details currently
are not fixed.