Prototyped one User view
Prototyped one User view as opposed to a Create and an Edit User view.
Will need to do this for the current data management items I have
already created but it makes more sense to do it this way (especially as
we add more data management items). No need to fix two GUIs when either
in Edit or Create mode.